Start progress towards gems being Matter.js objects
Started updating client and server code to use Matter.js world objects
in place of the basic JSON vectors that were previously representing the
gems. Currently nonfunctional as socket.io isn't able to emit an objects
on the scale of the Matter objects without a stack overflow. Will need
to work around this somehow.