Integrated into account based MVC framework from Air Hockey Online
Now uses AHO's MVC framework to allow for user accounts and sessions,
including storing user stats which are shown on the account page. Same
setup as before is working in place of the Air Hockey game (where
matching batch of physics objects is sent to each client and ran with
Matter through PIXI).