Some updates to server-client communication
Game manager now calculates its own delta time and uses this for
updating to help with server lag. Players compare time of received
object packets to the current time on the object and only update it if
the packet is newer than the most recent received packet. Player also
tries to instantly apply an update to objects upon receiving them to
simulate the time that's passed since the object's emission (i.e. to
compensate for the ping).