Begin prep for abilities and UI, broke up code a little, ported window-manager
Added window-manager module that was used in Leap of Faith and Lumina
repositories. Turns out the entire module seems to only use Victor for
the x,y data storage object and not for any vector math (which makes
sense with a UI engine), so it now uses a simple Vector(x, y) from the
utilities (no longer relying on Victor).

Set up ability enum and internal scrolling between them in the client.
InputManager was useless so set it up to accept and re-emit eventual
abilityChange events so clients can listen for them to know the other
user's current ability and display it in the interface.