Lots of cleanup and polish
Player spawned bodies have an image based on their currently selected
ability (the same as the icon). This syncs across clients. There are now
retaining walls on either side of the screen, and a couple of slanted
statics in the center. For some reason, angling the statics changes
their size, which is really annoying (matter why). Also, it appears
Matter doesn't do collision checks outside of the screen at all, which
makes sense but is annoying for using the retaining walls above the
screen.

Fixed the issue with clicks registering at the wrong location if the
screen is resized. Actually use the serverInfo element to display server
messages. Automatically hides itself again after 2 seconds.