Add more user feedback, more cleanup
Got rid of some random bits of unnecessary code, largely on the server
where there was still game code that wasn't needed at all anymore. Left
some temporarily in case we decide to do score checking and such on the
GameManager for some reason.

Players now draw a drag line as the prepare a shot, and always draw a
little orb where their cursor is. The enemy player can see their
orb/drag line as well, and their line color is based on the side of the
map they're on, matching the purple/orange outlines of the player walls.

Eventually we may wanna add a tag to objects to say who spawned them and
add a glow or something to indicate that (a colored glow)? That will
probably push us into the realm of needing a custom render loop, which
means taking time out to learn the Pixi documentation properly. Might be
worth it for the cool effects if we have the time.