Attempted fix for existing objects being overwritten
Would've been better to find out why it was happening in the first
place, the client player would sometimes sort of overwrite old objects
when they spawned in new ones, and it was because it was creating an
object with a new "unique" ID that was the same as an existing one.
Couldn't see a clear reason why that'd be the case so I now forcibly
assign them a new definitely unique ID before sending the processed obj
literal to the server as a request.

It seems possible for there to be some client-server desync again when
it comes to removing the body. This is probably because the host client
only syncs its own existing bodies. This means that if the client has a
body that the host does not, it will never get cleaned up because we do
no checks to see if objects shouldn't exist. Suggested fix is to
eventually have a per-player hash map cache on the server of minified
object data stored with object ID key and object data value. If a value
is set in the client hash map that doesn't exist in the host cache, the
server will tell the client player to remove it. The server can then
also act as a middleman for sending stuff to the client - if the
client's last sent information about a body matched the host's last sent
(and was recent enough to matter), it doesn't need to be emitted. Should
lower client-side traffic if we have time to implement that.