Added game ending and win/tie/loss stats Game can now end, disabling user interaction, world object spawning and emitting on the host, and sending user stats to the database. Users can view their stats on their account page.
It is possible to tie in this game, so this required adding gamesTied as an account variable on the Account Mongo schema. I'm not really sure how to count tied games for the win-loss ratio on the account page (as half a win, half a loss?). Currently it doesn't count them at all, and will still just compare the straight wins to straight losses. Which might be ok, but we'll need to change the text in the case of them having, for example, 1 win and 1 tie (You've won all 2 of your games played! isn't really true then).