Fixed issues with ending display and game stats being sent repeatedly
Fixes an issue where the forced counting of gems in the goal at the end
of the game would always visibly display as being counted for the host
regardless of whose goal they were in. This was just a visual error on
the HUD, internally the points and win/loss were being counted properly.

Fixed an issue with the resulting stats of a game being sent multiple
times. The server wasn't sending back a .json() on the res so it was
repeatedly sending the sendAjax call to try to make the connection. It
would then register the score for each attempt until the AJAX called
timed out entirely.