Cleaned up chat.js, added chat coloring based on team, fixed game length and other various bugs
Added a createMessage function to chat.js that creates a message in the
chat window, since we were repeating code all over the place for
appending children to create messages and stuff. File appears longer but
is actually notably shorter, it's just also more commented :v) Chat now
also colors based on your side and who your opponent is and stuff. I did
this with public variables because it was the easiest way to get the
engine IFFE to talk to the chat file. This creates "exploits" where the
player could edit the global variables via cmd line to change the chat
color, but that's honestly kind of cool.

Fixed game length being only 10 seconds for testing. Fixed sendAjax not
doing a redirect on the page if it receives one (e.g. refreshing account
page when changing username). You can send a redirect JSON of
"undefined" to not redirect (which happens when a game ends). Made
default "You are not logged in" chat message a server message, and made
it get replaced on connect so the first chat msg doesn't still say
you're not logged in even when you are. Made a documentation.jade
template, haven't started it.