Most of documentation, some style changes
Made documentation page, need to be logged in to access it. Not sure
with how our site is set up how to make it so it can be accessed while
logged in OR while logged out without it giving errors because of the
code in the base.jade. Tried to fix that for a while, but it ended up
giving Jade errors or making chat disabled on just that page, which was
awkward. Better to just have them log in to access the site
documentation.

Changed site styling so the sidebar is fixed to the left and scrolls
with the page. Mainly needed to because it wasn't following along on the
documentation page, which is long.